requirements.txt
Django==5.2.1
djangorestframework==3.14.0
django-cors-headers==4.9.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
